* 'develop' of ssh://stash.anoda.mobi:7999/zz/build-mirror:
  corrupted video fix
  minor fix

+++ Project/Zazo/Library/Layers/Service/ApplicationServices/ZZVideoFileHandler.m
@@ -319,7 +319,11 @@
                 [self deleteRemoteFileAndVideoId:video];
                 ZZVideoDomainModel* videoModel = [ZZVideoDataProvider modelFromEntity:video];
                 [ZZThumbnailGenerator generateThumbVideo:videoModel];
-                [self.delegate deleteAllViewedOrFailedVideosWithFriendId:friendID];
+                
+                if (![ZZThumbnailGenerator isVideoCorruptedWithModel:videoModel])
+                {
+                    [self.delegate deleteAllViewedOrFailedVideosWithFriendId:friendID];
+                }
                 
                 [self.delegate setAndNotifyIncomingVideoStatus:ZZVideoIncomingStatusDownloaded friendId:friendID videoId:videoId];
